+++ examples/token-stats/core/src/lib.rs
@@ -33,7 +33,7 @@ sol! {
 sol! {
     struct APRCommitment {
         Commitment commitment;
-        uint64 days;
+        uint64 numDays;
         uint64 finalBlockNumber;
         uint64 annualSupplyRate;
     }