+++ examples/token-stats/host/src/main.rs
@@ -102,7 +102,7 @@ async fn main() -> Result<()> {
 
     // Calculation is handling `/ 10^18 * 100` to match precision for a percentage.
     let apr = apr_commit.annualSupplyRate as f64 / 10f64.powi(16);
-    println!("Proven APR over {} days is: {apr}%", apr_commit.days);
+    println!("Proven APR over {} days is: {apr}%", apr_commit.numDays);
 
     Ok(())
 }